Arrhg! Someone deleted my turn_player fix. Allright. I'll do it again.

+++ src/sprite.c
@@ -43,8 +43,8 @@ short sample_sprite_color(double x, double y, sprite_t* texture) {
     assert(y >= 0);
     assert(y <= 1);
     assert(x <= 1);
-    int sprite_x = round(x * (texture->width - 1));
-    int sprite_y = round(y * (texture->height - 1));
+    int sprite_x = round(x * (double)(texture->width - 1));
+    int sprite_y = round(y * (double)(texture->height - 1));
     return get_sprite_color(sprite_x, sprite_y, texture);
 }
